Reorganize rust code for egui integration
No need for the wrapper or the render shard.
The `UI` shard doesn't return a context object anymore.
